requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.5
djangorestframework==3.14.0
idna==3.4
lxml==4.9.2
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.12
